Memory usage is reduced. Before, a product state was reprecented by a tuple. Now, a product state is instead represented by an integer.
Here is one product state example, expressed in the different formats:
(2, 5)
"0010010"
10
where there are two electrons and 7 spin-orbitals in total.
The tuple expresses which spin-orbitals are occupied by electrons.
The string shows the occupation of all spin-orbitals.
The integer is a compact format. It is the integer representation of the binary string expressed in base 2.

Using the integer representation instead of the tuple representation to store the product states should reduce the memory (RAM) usage, approximatly by a factor 20.
The memory usage is reduced in particular for product states containing many electrons.