Speedup with this version.
bitarrays are used for doing the product state manipulations (i.e. removing and adding electrons), instead of strings.
This gives a speedup.

bytes are used for storing a product state, instead of an integer. This takes about the same memory.
But it is faster to convert to bytes than to integer, starting from a bitarray representation.